story_refinement_loop_system/pyproject.toml
[project]
name = "story-refinement-loop-system"
version = "0.1.0"
description = "Example 5.1 loop workflow (writer + critic) built with ADK and uv."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "google-adk>=1.18.0",
  "google-genai>=0.3",
  "python-dotenv",
]

[tool.uv]
package = false
